Resolve timing issues
The original board had 2x 8MHz xtal and and 1x 21.477MHz xtal, so I don't know why 24.576MHz was specified here. The conflict in timing which bled into the reset timing for recovery and the t80 cpu timing into setup, as well as the stability of vsync in lower lag video modes, stems from this uneven division between of clk_sys and clk_main. Maybe @gaz68 knows why 24.576MHz was chosen instead of an even 24?